% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenescore.r
\name{mitogenescore}
\alias{mitogenescore}
\title{mitogenescore}
\usage{
data(mito_data)

data(mitol2)

res<-mitogenescore(mito_data,symbol=mitol2,method="ssgsea")

head(res[,1:6])
}
\arguments{
\item{mito_data}{transcriptome matrix}

\item{symbol}{mitocarta genesets by gene symbols: mitol1,mitol2,mitol3,mitolall}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}
}
\description{
mitogenescore
}
\examples{
data(mito_data)
data(mitol2)
res<-mitogenescore(mito_data,symbol=mitol2,method="ssgsea")
head(res[,1:6])
}
