% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoheat.r
\name{mitoheat}
\alias{mitoheat}
\title{mitoheat}
\usage{
data(mito_data)

data(mitol2)

data(mito_pheno)

res<-mitogenescore(mito_data,symbol=mitol2,method="ssgsea")

metaheat(res,pheno,scale="none",fontsize=10)
}
\arguments{
\item{results}{results obtained from mitogenescore function or sub-geneset}

\item{scale}{method of scaling "none" "row" "column"}

\item{fontsize}{size of the font on the graph}

\item{symbol}{phenotype file from tested samples}
}
\description{
mitoheat
}
\examples{
data(mito_data)
data(mitol2)
data(mito_pheno)
res<-mitogenescore(mito_data,symbol=mitol2,method="ssgsea")
mitoheat(res,mito_pheno,scale="none",fontsize=10)
}
